[FIX] logging in by username, added authorization in /collections resource

+++ src/main/java/com/blstream/myhoard/controller/CollectionController.java
@@ -3,11 +3,13 @@ package com.blstream.myhoard.controller;
 import com.blstream.myhoard.biz.exception.ErrorCode;
 import com.blstream.myhoard.biz.exception.MyHoardException;
 import com.blstream.myhoard.biz.model.CollectionDTO;
+import com.blstream.myhoard.biz.model.UserDTO;
 import com.blstream.myhoard.biz.service.ResourceService;
 import com.blstream.myhoard.db.model.SortResult;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
@@ -35,16 +37,19 @@ public class CollectionController {
      * Obsługa wypisywania/sortowania kolekcji
      * @param fieldName - po jakich polach sortować (domyślnie po nazwie)
      * @param sortDir   - w jakim kierunku ("asc" lub "desc"; domyślnie "asc")
+     * @param request
      * @return Lista kolekcji.
      */
     @RequestMapping(method = RequestMethod.GET)
     @ResponseStatus(HttpStatus.OK)
     @ResponseBody
     public List<CollectionDTO> getCollections(@RequestParam(value = "sort_by", defaultValue = "name") String[] fieldName,
-            @RequestParam(value = "sort_direction", defaultValue = "asc") String sortDir) {
+            @RequestParam(value = "sort_direction", defaultValue = "asc") String sortDir,HttpServletRequest request) {
+        UserDTO user =(UserDTO) request.getAttribute("user");
         Map<String, Object> params = new HashMap<>();
         params.put("sort_by", fieldName);
         params.put("sort_dir", sortDir);
+        params.put("username", user.getUsername());
         try {
             return collectionService.getList(params);
         } catch (Exception ex) {
@@ -75,9 +80,11 @@ public class CollectionController {
     @RequestMapping(method = RequestMethod.POST)
     @ResponseStatus(HttpStatus.CREATED)
     @ResponseBody
-    public CollectionDTO addCollection(@RequestBody @Valid CollectionDTO collection, BindingResult result) {
+    public CollectionDTO addCollection(@RequestBody @Valid CollectionDTO collection, BindingResult result,HttpServletRequest request) {
         if (result.hasErrors())
             throw new MyHoardException(400, result.getFieldError().getDefaultMessage());
+        UserDTO user = (UserDTO)request.getAttribute("user");
+        collection.setOwner(user.getUsername());
         collectionService.create(collection);
         return collection;
     }
@@ -85,9 +92,14 @@ public class CollectionController {
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     @ResponseStatus(HttpStatus.OK)
     @ResponseBody
-    public CollectionDTO getCollection(@PathVariable String id) {
+    public CollectionDTO getCollection(@PathVariable String id, HttpServletRequest request) {
         try {
-            return collectionService.get(Integer.parseInt(id));
+            UserDTO user = (UserDTO)request.getAttribute("user");
+            CollectionDTO collection = collectionService.get(Integer.parseInt(id));
+            if(collection.getOwner().equals(user.getUsername()))
+                return collection;
+            else
+                throw new MyHoardException(104,"Forbidden");
         } catch (NumberFormatException ex) {
             throw new MyHoardException(300, "Niepoprawne id: " + id);
         }
@@ -97,19 +109,31 @@ public class CollectionController {
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
     @ResponseStatus(HttpStatus.OK)
     @ResponseBody
-    public CollectionDTO updateCollection(@PathVariable String id, @Valid @RequestBody CollectionDTO collection, BindingResult result) {
+    public CollectionDTO updateCollection(@PathVariable String id, @Valid @RequestBody CollectionDTO collection,HttpServletRequest request, BindingResult result) {
         if (collection.getName() != null && result.hasErrors())
             throw new MyHoardException(320, result.getFieldError().getDefaultMessage());
-        collection.setId(id);
-        collectionService.update(collection);
-        return collection;
+        UserDTO user = (UserDTO)request.getAttribute("user");
+        CollectionDTO tmp = collectionService.get(Integer.parseInt(id));
+        if(tmp.getOwner().equals(user.getUsername())) {
+            collection.setId(id);
+            collectionService.update(collection);
+            return collection;
+        } else {
+            throw new MyHoardException(103,"Forbidden");
+        }
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void removeCollection(@PathVariable String id) {
+    public void removeCollection(@PathVariable String id, HttpServletRequest request) {
         try {
-            collectionService.remove(Integer.parseInt(id));
+            UserDTO user = (UserDTO)request.getAttribute("user");
+            CollectionDTO tmp = collectionService.get(Integer.parseInt(id));
+            if(tmp.getOwner().equals(user.getUsername())) {
+                collectionService.remove(Integer.parseInt(id));
+            } else {
+                throw new MyHoardException(103,"Forbidden");
+            }
         } catch (NumberFormatException ex) {
             throw new MyHoardException(400, "Niepoprawne id: " + id);
         }
